backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt==4.0.1",
    "email-validator>=2.3.0",
    "fastapi>=0.135.3",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "uvicorn[standard]>=0.43.0",
]
